+++ packages/gafl-webapp-service/src/plugins.js
@@ -1,44 +1,19 @@
-import db from 'debug'
 import Inert from '@hapi/inert'
 import Vision from '@hapi/vision'
 import Disinfect from 'disinfect'
 import Scooter from '@hapi/scooter'
 import Blankie from 'blankie'
 import Crumb from '@hapi/crumb'
-import HapiGapi from '@defra/hapi-gapi'
 import Cookie from '@hapi/cookie'
 import HapiI18n from 'hapi-i18n'
-import { useSessionCookie } from './session-cache/session-manager.js'
 import { getCsrfTokenCookieName } from './server.js'
-import { trackAnalyticsAccepted, getAnalyticsSessionId, pageOmitted } from '../src/handlers/analytics-handler.js'
 import Dirname from '../dirname.cjs'
 import path from 'path'
-
-const debug = db('webapp:plugin')
 
 // This is a hash provided by the GOV.UK Frontend:
 // https://frontend.design-system.service.gov.uk/importing-css-assets-and-javascript/#use-a-hash-to-unblock-inline-javascript
 // It is added to the CSP to except the in-line script. It needs the quotes.
 const scriptHash = "'sha256-GUQ5ad8JK5KmEWmROf3LZd9ge94daqNvd8xy9YS1iDw='"
-
-const trackAnalytics = async request => {
-  const pageOmit = await pageOmitted(request)
-  const canTrack = await trackAnalyticsAccepted(request, pageOmit)
-  const optDebug = process.env.ENABLE_ANALYTICS_OPT_IN_DEBUGGING?.toLowerCase() === 'true'
-  if (optDebug) {
-    const sessionId = await getAnalyticsSessionId(request)
-    if (canTrack === true) {
-      debug(`Session is being tracked for: ${sessionId}`)
-    } else {
-      if (pageOmit === true) {
-        debug(`Session is not tracking current page for: ${sessionId}`)
-      } else {
-        debug(`Session is not being tracked for: ${sessionId}`)
-      }
-    }
-  }
-  return canTrack
-}
 
 const initialiseDisinfectPlugin = () => ({
   plugin: Disinfect,
@@ -49,6 +24,9 @@ const initialiseDisinfectPlugin = () => ({
   }
 })
 
+const unsafeInline = 'unsafe-inline'
+const googleTagUrl = '*.googletagmanager.com'
+
 const initialiseBlankiePlugin = () => ({
   plugin: Blankie,
   options: {
@@ -57,10 +35,21 @@ const initialiseBlankiePlugin = () => ({
      * It must allow web-fonts from 'fonts.gstatic.com'
      */
     fontSrc: ['self', 'fonts.gstatic.com', 'data:'],
-    scriptSrc: ['self', 'unsafe-inline', scriptHash, 'www.googletagmanager.com'],
-    connectSrc: ['self', '*.google-analytics.com'],
+    scriptSrc: [
+      'self',
+      unsafeInline,
+      scriptHash,
+      googleTagUrl,
+      '*.tagassistant.google.com',
+      'unsafe-eval',
+      'https://tagmanager.google.com/'
+    ],
+    connectSrc: ['self', '*.google-analytics.com', googleTagUrl, '*.analytics.google.com'],
     generateNonces: true,
-    frameAncestors: 'none'
+    frameAncestors: 'none',
+    imgSrc: ['self', unsafeInline, googleTagUrl, 'fonts.gstatic.com', 'data:', 'https://ssl.gstatic.com/'],
+    manifestSrc: ['self'],
+    styleSrc: [unsafeInline, 'self', 'fonts.googleapis.com', googleTagUrl, 'https://tagmanager.google.com/']
   }
 })
 
@@ -75,35 +64,6 @@ const initialiseCrumbPlugin = () => ({
     logUnauthorized: true
   }
 })
-
-const initialiseHapiGapiPlugin = () => {
-  const hapiGapiPropertySettings = []
-  if (process.env.ANALYTICS_PRIMARY_PROPERTY) {
-    hapiGapiPropertySettings.push({
-      id: process.env.ANALYTICS_PRIMARY_PROPERTY,
-      key: process.env.ANALYTICS_PROPERTY_API,
-      hitTypes: ['page_view']
-    })
-  } else {
-    console.warn("ANALYTICS_PRIMARY_PROPERTY not set, so Google Analytics won't track this")
-  }
-
-  return {
-    plugin: HapiGapi,
-    options: {
-      propertySettings: hapiGapiPropertySettings,
-      trackAnalytics: trackAnalytics,
-      sessionIdProducer: async request => {
-        let sessionId = null
-        if (useSessionCookie(request)) {
-          const { gaClientId } = await request.cache().helpers.status.get()
-          sessionId = gaClientId ?? (await request.cache().getId())
-        }
-        return sessionId
-      }
-    }
-  }
-}
 
 const initialiseHapiI18nPlugin = () => {
   const showWelshContent = process.env.SHOW_WELSH_CONTENT?.toLowerCase() === 'true'
@@ -126,7 +86,6 @@ export const getPlugins = () => {
     initialiseDisinfectPlugin(),
     initialiseBlankiePlugin(),
     initialiseCrumbPlugin(),
-    initialiseHapiGapiPlugin(),
     initialiseHapiI18nPlugin()
   ]
 }