+++ packages/gafl-webapp-service/src/services/payment/__test__/govuk-pay-service.spec.js
@@ -55,7 +55,8 @@ describe('The govuk-pay-service', () => {
           i18n: {
             getCatalog: () => ({
               over_66: ' (Over 66)',
-              licence_type_radio_salmon: 'Salmon and sea trout'
+              licence_type_radio_salmon: 'Salmon and sea trout',
+              licence_type_12m: '12 months'
             })
           },
           info: { host: '0.0.0.0:3000' },
@@ -97,7 +98,8 @@ describe('The govuk-pay-service', () => {
           i18n: {
             getCatalog: () => ({
               over_66: ' (Over 66)',
-              licence_type_radio_salmon: 'Salmon and sea trout'
+              licence_type_radio_salmon: 'Salmon and sea trout',
+              licence_type_8d: '8 days'
             })
           },
           info: { host: '0.0.0.0:3000' },